requirements.txt
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1.1
django-cors-headers==4.0.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2